feat: update sql runner config layout
