+++ packages/frontend/src/components/DataViz/config/CartesianChartConfiguration.tsx
@@ -2,8 +2,9 @@ import { ChartKind, type VizColumn } from '@lightdash/common';
 import { Stack, Tabs } from '@mantine/core';
 import { barChartConfigSlice } from '../store/barChartSlice';
 import { lineChartConfigSlice } from '../store/lineChartSlice';
+import { CartesianChartAxes } from './CartesianChartAxes';
 import { CartesianChartFieldConfiguration } from './CartesianChartFieldConfiguration';
-import { CartesianChartStyling } from './CartesianChartStyling';
+import { CartesianChartSeries } from './CartesianChartSeries';
 
 export const CartesianChartConfig = ({
     selectedChartType,
@@ -29,7 +30,7 @@ export const CartesianChartConfig = ({
                 <Tabs.List>
                     <Tabs.Tab value="data">Data</Tabs.Tab>
                     <Tabs.Tab value="series">Series</Tabs.Tab>
-                    <Tabs.Tab value="display">Display</Tabs.Tab>
+                    <Tabs.Tab value="axes">Axes</Tabs.Tab>
                 </Tabs.List>
 
                 <Tabs.Panel value="data" pt="xs">
@@ -41,14 +42,14 @@ export const CartesianChartConfig = ({
                 </Tabs.Panel>
 
                 <Tabs.Panel value="series" pt="xs">
-                    <CartesianChartStyling
+                    <CartesianChartSeries
                         actions={actions}
                         selectedChartType={selectedChartType}
                     />
                 </Tabs.Panel>
 
-                <Tabs.Panel value="display" pt="xs">
-                    <CartesianChartStyling
+                <Tabs.Panel value="axes" pt="xs">
+                    <CartesianChartAxes
                         actions={actions}
                         selectedChartType={selectedChartType}
                     />
