+++ packages/frontend/src/features/semanticViewer/components/SemanticViewerVizConfig/CartesianVizStyling.tsx
@@ -1,5 +1,5 @@
-import { CartesianChartStyling } from '../../../../components/DataViz/config/CartesianChartStyling';
+import { CartesianChartAxes } from '../../../../components/DataViz/config/CartesianChartAxes';
 
 // NOTE: This file is included here, but the styling is shared from the
 // data viz folder for now. If the styling diverges, we can add it here.
-export { CartesianChartStyling as CartesianVizStyling };
+export { CartesianChartAxes as CartesianVizStyling };
